// user-management-module-front/src/app/models/view/subscription-view.ts
export interface SubscriptionViewInfo {
    id: string;
    companyName: string;
    zipCode: string;
    address: string;
    contactNo: string;
    subscriptionEmail: string;
    companyCategory: Array<string>;
    isHuman: boolean;
    isAgree: boolean;
    response: string;
    backendDate: string;
    frontendDate: string;
    subscriberName: string,
    subscriberMail: string
}